# Optimize series key parsing on startup