###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass
- [x] CHANGELOG.md updated

This optimizes the `ParseKey` func to help speed up startup times for shards with many keys or very long keys. This should help #6250 in some cases.  `ParseKey` is called for every key when reloading the WAL and TSM file indexes.  The main change is avoiding calling the unescape funcs for tag keys and values which always returned a new string even if there was nothing escaped.  We also quickly scan the key to see if we might need to unescape anything and skip all of those calls if we don't need to.  The remaining allocation time is creating the tags map.  The tags map is really just a temporary holder to allow the tags to be indexed.  We could rework this further to avoid that map altogether and parse the key directly into the index structures.

```
benchmark               old ns/op     new ns/op     delta
BenchmarkParseKey-8     5621          2050          -63.53%

benchmark               old allocs     new allocs     delta
BenchmarkParseKey-8     25             24             -4.00%

benchmark               old bytes     new bytes     delta
BenchmarkParseKey-8     1318          1030          -21.85%
```