@daviesalex 2016-05-28
We just moved our server (~2T of InfluxDB data) to the nightly from last night (with this in it) vs the nightly from the 23rd May. The startup time goes from over an hour to just over 10 minutes. The system load during the startup is roughly doubled (though there are still boat loads of CPUs doing nothing). The output of "perf top" shows totally different limiting factors. 

Major CPU users now:

```
   7.37%  influxd                       [.] runtime.scanobject
   6.02%  influxd                       [.] runtime.heapBitsForObject
   5.57%  [kernel]                      [k] _raw_spin_lock
   5.51%  influxd                       [.] runtime.greyobject
   4.72%  influxd                       [.] sync.(*RWMutex).Lock
   4.55%  influxd                       [.] runtime.mallocgc
   4.14%  influxd                       [.] runtime.heapBitsSweepSpan
   3.78%  influxd                       [.] runtime.mapassign1
   3.37%  influxd                       [.] type..hash.[8]github.com/influxdata/influxdb/monitor/diagnostics.Client
```

This is a huge win. The next win is to just do more of startup in different goroutines, but thats a totally different kettle of fish.

Thank you!!